Crate filler_wasm_code

A Filler calulation algorithm.

Long story short, this is an implementation of the minimax algorithm applied to the game Filler, which is present in the iMessage GamePigeon app. So far, I have implemented alpha-beta pruning and transposition tables, and I am planning on implementing pre-sorting soon (which works well with alpha-beta pruning).

Below is a list of everything exposed to JavaScript:

gen_random_board() - See prep module

get_choices() - See decision_making module

solve_board() - See below

Board struct - See board module

from_str()

get_string()

get_cell()

get_player_turn()

get_player_squares()

run_turn()

Modules

board

Describes a board object which contains both an 8x7 board, as well as the current player whose turn it is to move. This file also includes utility functions to convert board cells from chars to ints and vice versa.

cache

This file describes 2 different caches, an information cache and a value cache. These caches are used to reduce redundant calculations, as some boards are encountered more than once throughout a search.

decision_making

This file implements the minimax algorithm, applied to the Filler game.

distribution_timer

The DistributionTimer is used to time how long each section of a given algorithm takes, specifically for the minimax algorithm.

prep

This file is used for any data preparation necessary. Right now, it just allows you to generate a random board.

stat_calculation

These functions are used for calculations for means and standard deviations Nothing here is WASM-compatible.

stats

The Stats struct is used to capture various statistics generated during the runtime of the algorithm. This includes: