Module filler_wasm_code::board

Describes a board object which contains both an 8x7 board, as well as the current player whose turn it is to move. This file also includes utility functions to convert board cells from chars to ints and vice versa.

Here is the relationship between colors, chars, and the corresponding int.

Color	Char	Int

Red	R	0

Green	G	1

Yellow	Y	2

Blue	U	3

Purple	P	4

Black	K	5

The following functions are exposed to Javascript:

from_str()

get_player_turn()

get_string()

get_cell()

run_turn()

get_player_squares()

Structs

Board

A Board struct keeping track of the current state, specifically the current board and the current player’s turn

Functions

get_char_from_int

Converts an integer to its corresponding char

get_int_from_char

Converts a char to its corresponding integer