Struct filler_wasm_code::board::Board

pub struct Board { /* private fields */ }

A Board struct keeping track of the current state, specifically the current board and the current player’s turn

Implementations

impl Board

Everything here is not visible to the JavaScript

pub fn new(board: [u8; 56], player_turn: u8) -> Board

Creates a new board from the current player turn and an array of 56 ints, corresponding to each cell of the board in row-column order

Column 0	Column 1	Column 2	Column 3	Column 4	Column 5	Column 6	Column 7

0	1	2	3	4	5	6	7

16	17	18	19	20	21	22	23

24	25	26	27	28	29	30	31

32	33	34	35	36	37	38	39

40	41	42	43	44	45	46	47

48	49	50	51	52	53	54	55

pub fn new_not_flat(board: [[u8; 8]; 7], player_turn: u8) -> Board

Creates a new board from the current player turn and an array of 56 ints, corresponding to each cell of the board in row-column order

Column 0	Column 1	Column 2	Column 3	Column 4	Column 5	Column 6	Column 7

0	1	2	3	4	5	6	7

16	17	18	19	20	21	22	23

24	25	26	27	28	29	30	31

32	33	34	35	36	37	38	39

40	41	42	43	44	45	46	47

48	49	50	51	52	53	54	55

pub fn get_board(&self) -> &[[u8; 8]; 7]

Returns an immutable reference to the private board

pub fn get_player_claimable(&self, player_number: u8) -> Vec<u8>

Returns a list of all the different colors that a player can choose to claim new square(s)

pub fn get_player_claimable_from_squares(
&self,
player_number: u8,
squares: &Vec<u8>
) -> Vec<u8>

Returns a list of all the different colors that a player can choose to claim new square(s), with a list of all the squares owned by that player to speed up this operation

pub fn run_turn_from_squares(&self, choice: u8, squares: &Vec<u8>) -> Board

Changes all of the player’s squares from their current color to the new choice, given that choice as well as a list of the cell indices that player owns

pub fn display_board(&self)

Prints out the current board, complete with ANSI escape codes for coloring

impl Board

Everything here is visible to the JavaScript

pub fn from_str(board: &str) -> Board

Creates a board from a string (of length 56), composed of the chars in order of each cell of the board, as well as the current player’s turn

pub fn get_player_turn(&self) -> u8

Returns an immutable reference to the private player turn

pub fn get_string(&self) -> String

Returns a string describing the current board, which can be useful for copying the board and storing it

pub fn get_cell(&self, row: usize, column: usize) -> u8

Returns the cell at the given row and column

pub fn run_turn(&self, choice: u8) -> Board

Returns a new board after the current player chooses the given color and the board is updated accordingly

pub fn get_player_squares(&self, player_number: u8) -> Vec<u8>

Calculates which board indices belong to the given player