Module filler_wasm_code::cache

This file describes 2 different caches, an information cache and a value cache. These caches are used to reduce redundant calculations, as some boards are encountered more than once throughout a search.

Each cache is a HashMap, one storing informational data (stored as a “StoredBoardInfo”), and one storing valued data (stored as a “StoredBoardValue”). StoredBoardInfo contains a list of which board indices belong to a given player. StoredBoardValue contains which choices a player can make, the score value of the current board, and the depth at which this score was calculated.

StoredBoardInfo and StoredBoardValue are stored separately because while StoredBoardInfo will always be valid, StoredBoardValue will not always be valid, so time and space can be saved over the course of an iteratively deepening search by clearing the stored values at each depth, but saving the stored information across depths.

As of now, none of this is exposed to JavaScript, as it would significantly increase the memory footprint of each cache if these structs had their own getter functions, and JavaScript does not support public, immutable arrays. In the future, a HashMap manager can be created to manage HashMap interaction if this information should be exposed to JavaScript.

Structs

StoredBoardInfo

The StoredBoardInfo contains an array of cell indices that belong to each player

StoredBoardValue

The StoredBoardValue contains the valid choices for each player, and the score of the given board at the given depth

Functions

array_to_vec_6

This helper function converts a stored array of choices to the associated Vec. It removes the padding of 6’s.

array_to_vec_56

This helper function converts a stored array of cell indices to the associated Vec. It removes the padding of 56’s.

push_infomap

Pushes the given information to the given information hash map

push_valuemap

Pushes the given information to the given value hash map

vec_to_array_6

This helper function converts a Vec of choices to the array they are stored as in the cache. The array is padded with 6’s to the maximum length.

vec_to_array_56

This helper function converts a Vec of cells indices to the array they are stored as in the cache. The array is padded with 56’s to the maximum length.