Module filler_wasm_code::decision_making

This file implements the minimax algorithm, applied to the Filler game.

The algorithm times its runtime distribution across each section using a DistributionTimer. The public implementations of each function wraps underlying functions that handle the actual computation required by the minimax algorithm.

The following functions are available to JavaScript:

get_choices()

Hopefully get_best_move and get_move_value will be made available to JavaScript in the future, once HashMap manipulation is given some wrapper functions.

Functions

get_best_move

This function will return a Vec of the various choices and associated scores calculated for those choices for the given board. It requires a board to calculate on, the maximum depth to search, a hashmap to use as a information cache, a hashmap to use as a value cache (for more information, see cache.rs), a stats object to store various statistics, and a distribution timer to store how much time is spent in each phase of the algorithm (to be used for later optimization)

get_choices

This function returns a list of all the possible, useful choices a player has to choose between for their turn