Struct filler_wasm_code::stats::Stats

pub struct Stats {
    pub ab_prunes: usize,
    pub boards_checked: usize,
    pub cache_accesses: usize,
    pub cache_duplicates: usize,
}

For an explanation of what each field represents, see the comment at the top of the function

Fields
ab_prunes: usizeboards_checked: usizecache_accesses: usizecache_duplicates: usize
Implementations

impl Stats

pub fn new() -> Stats

pub fn combine(&mut self, other_stats: &Stats)